requirements.txt
click==6.7
Flask==0.12.2
grpcio==1.11.0
grpcio-tools==1.11.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
protobuf==3.5.2.post1
six==1.11.0
Werkzeug==0.14.1
